refactor: 一旦useLifelines消してuseQuizに集約
